add interface for dropdown options
